minor fixes and tests fixes

+++ apps/server/src/modules/tldraw/service/tldraw-files-storage.service.ts
@@ -10,6 +10,10 @@ export class TldrawFilesStorageAdapterService {
 		const fileRecords = await this.filesStorageClientAdapterService.listFilesOfParent(docName);
 		const fileRecordIdsForDeletion = this.foundAssetsForDeletion(fileRecords, usedAssets);
 
+		if (fileRecordIdsForDeletion.length === 0) {
+			return;
+		}
+
 		await this.filesStorageClientAdapterService.deleteFiles(fileRecordIdsForDeletion);
 	}
